Move plugin files to ee_plugin module
Clean up installed plugin to include only extlibs and python files.
