+++ ee_plugin/metadata.txt
